kick-resume/requirements.txt
Flask==2.2.2
pdfminer.six==20221105
python-mammoth==1.4.1
google-generativeai==0.8.0
python-dotenv==1.0.1
Flask-Cors==4.0.1
openai-agents==0.0.6 # Assuming a recent version, adjust if needed
pydantic==2.8.2 # Assuming a recent version, adjust if needed
